EXPERIMENTAL - look into running provider outside k8s cluster
This commit marks investigation used to determine if the provider
could exist outside the k8s cluster such that only the kernels run
in the cluster.

This appears to be doable (and works with these changes for vanilla
kernels).  We would need to convey the service host and port used
for the SPARK_OPTS master url for spark kernels. This appears to be
achievable using `client.CoreApi().get_api_versions()` which returns
the following...
```
{'api_version': None,
 'kind': 'APIVersions',
 'server_address_by_client_cid_rs': [{'client_cidr': '0.0.0.0/0',
                                      'server_address': '192.168.65.3:6443'}],
 'versions': ['v1']}
```
so ret.server_address_by_client_cid_rs[0].server_address would likely
be the desired value.

However, after discussing with team, we feel the security implications
may be too great for this since it would require that kube-config
allow for things like namespace/pod creation/deletion, etc.  That
said, it "might" be something someone would want to allow some day.

Note that this commit includes the launcher locally, which has not
been done in the master branch at this time (but probably will be).
Care should be taken to properly merge this commit should it be revived.